Payment Added along with Subject Choices

+++ database/migrations/2021_10_19_164148_create_transactions_table.php
@@ -16,9 +16,9 @@ class CreateTransactionsTable extends Migration
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
             $table->foreignId('student_id')->constrained('students','id');
-            $table->integer('amount_due');
-            $table->integer('amount_paid');
-            $table->integer('balance_amt');
+            $table->decimal('amount_due');
+            $table->decimal('amount_paid');
+            $table->decimal('balance_amt');
             $table->year('year_of_exam');
         });
     }
